fixed issue with i18n support as per issue #45, removed demo files, relocated stories
