--- src/stories/Header.stories.svelte
@@ -1,32 +1,0 @@
-<script>
-import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
-import Header from './Header.svelte';
-</script>
-
-<Meta
-  title="Example/Header"
-  component="{Header}"
-  argTypes="{{
-    onLogin: { action: 'onLogin' },
-    onLogout: { action: 'onLogout' },
-    onCreateAccount: { action: 'onCreateAccount' },
-  }}"
-/>
-
-<Template let:args>
-  <Header
-    {...args}
-    on:login="{args.onLogin}"
-    on:logout="{args.onLogout}"
-    on:createAccount="{args.onCreateAccount}"
-  />
-</Template>
-
-<Story
-  name="LoggedIn"
-  args="{{
-    user: {},
-  }}"
-/>
-
-<Story name="LoggedOut" args="{{}}" />
