+++ src/stories/elements/GasCard.stories.svelte
@@ -1,12 +1,15 @@
 <script>
 import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
-import GasCard from './GasCard.svelte';
+import LocaleWrapper from '../LocaleWrapper.svelte';
+import GasCard from '../../components/elements/GasCard.svelte';
 </script>
 
 <Meta title="Elements/GasCard" component="{GasCard}" />
 
 <Template let:args>
-  <GasCard {...args} />
+  <LocaleWrapper>
+    <GasCard {...args} />
+  </LocaleWrapper>
 </Template>
 
 <Story name="GasCard" />
